+++ src/preload.ts
@@ -1,4 +1,5 @@
 import { Config } from "./plugins/config";
+import { Search } from "./plugins/search";
 
 import { IExports, IPlugin } from "./types/utools";
 
@@ -6,6 +7,7 @@ const globalThis = window as unknown as { exports: IExports, utools: UToolsApi, 
 
 const plugins: IPlugin[] = [
     new Config('config'),
+    new Search('search'),
 ];
 
 const exports: IExports = {};